changes to get tests to work with %2F
