+++ src/protagonist/DLCS.Repository/NamedQueries/Parsing/StoredNamedQueryParser.cs
@@ -59,10 +59,7 @@ public abstract class StoredNamedQueryParser<T> : BaseNamedQueryParser<T>
         var key = GetTemplateFromSettings(namedQuerySettings)
             .Replace("{customer}", parsedNamedQuery.Customer.ToString())
             .Replace("{queryname}", parsedNamedQuery.NamedQueryName)
-            .Replace("{args}",
-                string.Join("/",
-                    parsedNamedQuery.Args.Select(
-                        x => x.Replace(PathReplacement, "/", StringComparison.OrdinalIgnoreCase))));
+            .Replace("{args}", string.Join("/", parsedNamedQuery.Args));
 
         if (parsedNamedQuery.ObjectName.HasText()) key += $"/{parsedNamedQuery.ObjectName}";
         if (isControlFile) key += ".json";
